+++ CHANGELOG.md
@@ -18,11 +18,27 @@ None
 
 _This section contains implemented but not released features._
 
+## [0.2.0] - 2024-03-19
+
+### Added
+
+- `getNetworks` and `getStartupWelcomeText` functions
+
+### Changed
+
+- Update `@cinnabar-forge/eslint-plugin` to v0.4.0
+
+### Removed
+
+- `getNetworkAddressesList` function
+- `eslint-plugin-jsdoc` npm package
+
 ## [0.1.0] - 2024-01-28
 
 ### Added
 
-- Add `getNetworkAddressesList` function.
+- `getNetworkAddressesList` function.
 
-[unreleased]: https://github.com/cinnabar-forge/node-utils/compare/v0.1.0...HEAD
+[unreleased]: https://github.com/cinnabar-forge/node-utils/compare/v0.2.0...HEAD
+[0.2.0]: https://github.com/cinnabar-forge/node-utils/releases/tag/v0.2.0
 [0.1.0]: https://github.com/cinnabar-forge/node-utils/releases/tag/v0.1.0